Attn, Cable Loss measurement and Database Management Completed

+++ server/resultsDB/models.go
@@ -13,7 +13,7 @@ type CableLoss struct {
 	Date        string
 	Time        string
 	CableName   string
-	CableLength int64
+	CableLength float64
 	Loss        string
 }
